Move breakpoints.js to Constants folder, add comment to variables.scss about matching breakpoint.js
